Refactor the previous working solution to simpler solution
In the last solution I used the array data structure to keep track
and finding the last item that got push in the stack so I can
pop the item.

In the current solution I use a variable to keep track of the
last item that got push on the stack. Now I can simply use
that variable to pop the item out of the stack.

This solution save me from using multiple operations on the array
to find the stack's last position, which helps simplify the solution